Refactor with routers and update stats page
